Added profile switching command
Added a command that can be used to switch profiles,
Refactored yeochan's profile class names/file structure

+++ abstractConnection/src/com/yoke/connection/messages/app/OpenProfileCmd.java
@@ -8,13 +8,13 @@ public class OpenProfileCmd extends AppCmd {
     private static final long serialVersionUID = 4298204355886298903L;
 
     // The ID of the profile to be opened
-    public int profileID;
+    public long profileID;
 
     /**
      * Creates a new open profile command, to open a specific profile
      * @param profileID  The ID of the profile to open
      */
-    public OpenProfileCmd(int profileID){
+    public OpenProfileCmd(long profileID){
         this.profileID = profileID;
     }
 }
